+++ app/src/docs/_examples/pickerInput/CascadeSelectionModes.example.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import { DataPickerRow, FlexRow, MultiSwitch, FlexCell, PickerInput, PickerItem } from '@epam/uui';
-import { DataQueryFilter, DataRowProps, DataSourceState, useLazyDataSource, useUuiContext } from '@epam/uui-core';
+import { FlexRow, MultiSwitch, FlexCell, PickerInput } from '@epam/uui';
+import { DataQueryFilter, useLazyDataSource, useUuiContext } from '@epam/uui-core';
 import { Location } from '@epam/uui-docs';
 
 const cascadeSelectionModes: Array<{ id: 'explicit' | 'implicit'; caption: string }> = [
@@ -29,19 +29,9 @@ export default function CascadeSelectionModesExample() {
             getId: (i) => i.id,
             getParentId: (i) => i.parentId,
             getChildCount: (l) => l.childCount,
-            flattenSearchResults: true,
         },
         [],
     );
-
-    const getSubtitle = ({ path }: DataRowProps<Location, string>, { search }: DataSourceState) => {
-        if (!search) return;
-
-        return path
-            .map(({ value: pathValue }) => pathValue?.name)
-            .filter(Boolean)
-            .join(' / ');
-    };
 
     return (
         <FlexCell width={ 350 }>
@@ -57,14 +47,6 @@ export default function CascadeSelectionModesExample() {
                 selectionMode="multi"
                 valueType="id"
                 cascadeSelection={ cascadeSelection }
-                renderRow={ (props: DataRowProps<Location, string>, dataSourceState) => (
-                    <DataPickerRow
-                        { ...props }
-                        key={ props.rowKey }
-                        padding="12"
-                        renderItem={ (item) => <PickerItem { ...props } title={ item.name } subtitle={ getSubtitle(props, dataSourceState) } /> }
-                    />
-                ) }
             />
         </FlexCell>
     );